+++ src/components/Footer/ArrowUp.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {RiArrowUpSLine} from 'react-icons/ri';
+
+const ArrowUp = () => {
+  return (
+    <>
+      <a href="#header-container" className="arrow-up"><RiArrowUpSLine /></a>
+    </>
+  )
+}
+
+export default ArrowUp;
