+++ src/components/contact/Contact.js
@@ -1,6 +1,6 @@
 import React from 'react'; 
 import aboutData from "../about/data";
-import Footer from "./Footer";
+import Footer from "../Footer/Footer";
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import { makeStyles } from '@material-ui/core/styles';
